Dialogs are now managed by GraphEditor
