+++ client/src/app/architecture/architecture-editing/architecture-editing.service.ts
@@ -4,9 +4,7 @@ import { AddDatastoreCommand, AddMessageBrokerCommand, AddMessageRouterCommand, 
 import { ConfirmationService, MessageService } from 'primeng/api';
 import { DialogService } from 'primeng/dynamicdialog';
 import { GraphInvoker } from '../../commands/invoker';
-import { SessionService } from '../../core/session/session.service';
 import { GraphService } from '../../graph/graph.service';
-import { DialogAddNodeComponent } from '../dialog-add-node/dialog-add-node.component';
 
 import { g } from 'jointjs';
 import { AddLinkCommand, RemoveLinkCommand } from '../../commands/link-commands';
@@ -70,91 +68,15 @@ export class ArchitectureEditingService {
   }
 
   deleteNode(node) {
-    this.confirmationService.confirm({
-        message: 'Do you want to delete this node?',
-        header: 'Node Deletion Confirmation',
-        icon: 'pi pi-exclamation-triangle',
-        accept: () => {
-            this.graphInvoker.executeCommand(new RemoveNodeCommand(this.graphService.getGraph(), node))
-            this.messageService.add({ severity: 'success', summary: 'Confirmed', detail: `Node ${node.getName()} deleted succesfully` });
-        },
-        reject: () => {
-            this.messageService.add({ severity: 'info', summary: 'Rejected', detail: `Node ${node.getName()} not deleted` });
-        }
-    });
+    this.graphInvoker.executeCommand(new RemoveNodeCommand(this.graphService.getGraph(), node));
   }
 
-  deleteService(node) {
-    this.confirmationService.confirm({
-        message: 'Do you want to delete this service?',
-        header: 'Node Deletion Confirmation',
-        icon: 'pi pi-exclamation-triangle',
-        accept: () => {
-            // this.graphInvoker.executeCommand(new RemoveNodeCommand(this.gs.getGraph(), node))
-            this.graphInvoker.executeCommand(new RemoveServiceCommand(this.graphService.getGraph(), node.getName()));
-        },
-        reject: () => {
-            node.hideIcons();
-            this.messageService.add({ severity: 'info', summary: 'Rejected', detail: `Node ${node.getName()} not deleted` });
-        }
-    });
-  }
-
-  deleteDatastore(node) {
-    this.confirmationService.confirm({
-      message: 'Do you want to delete this datastore?',
-      header: 'Node Deletion Confirmation',
-      icon: 'pi pi-exclamation-triangle',
-      accept: () => {
-          // this.graphInvoker.executeCommand(new RemoveNodeCommand(this.gs.getGraph(), node))
-          this.graphInvoker.executeCommand(new RemoveDatastoreCommand(this.graphService.getGraph(), node.getName()));
-      },
-      reject: () => {
-          node.hideIcons();
-          this.messageService.add({ severity: 'info', summary: 'Rejected', detail: `Datastore ${node.getName()} not deleted` });
-      }
-  });
-  }
-
-  removeLink(link: joint.shapes.microtosca.RunTimeLink) {
-    console.log("removing link called");
-    console.log(this.confirmationService);
-    this.confirmationService.confirm({
-        message: 'Do you want to delete the link?',
-        header: 'Link deletion',
-        icon: 'pi pi-exclamation-triangle',
-        accept: () => {
-            this.graphInvoker.executeCommand(new RemoveLinkCommand(this.graphService.getGraph(), link));
-            //this.messageService.add({ severity: 'success', summary: 'Confirmed', detail: `Node ${node.getName()} deleted succesfully` });
-        },
-        reject: () => {
-            this.messageService.add({ severity: 'info', summary: 'Rejected', detail: `Link not deleted` });
-        }
-    });
-
-  }
-
-  removeCommunicationPattern(node) {
-    this.graphInvoker.executeCommand(new RemoveCommunicationPatternCommand(this.graphService.getGraph(), node.getName()));
-  }
-
-  addLink(leftClickSelectedNode, node) {
-    console.log("selected, new ->", leftClickSelectedNode, node);
-    const ref = this.dialogService.open(DialogAddLinkComponent, {
-        data: {
-            source: leftClickSelectedNode,
-            target: node
-        },
-        header: 'Add a link',
-    });
-    ref.onClose.subscribe((data) => {
-        if (data) {
-            var command = new AddLinkCommand(this.graphService.getGraph(), leftClickSelectedNode.getName(), node.getName(), data.timeout, data.circuit_breaker, data.dynamic_discovery);
-            this.graphInvoker.executeCommand(command);
-            this.navigation.getPaper().findViewByModel(leftClickSelectedNode).unhighlight();
-            console.log("added link");
-        }
-    });
+  addLink(leftClickSelectedNode, target, timeout?, circuit_breaker?, dynamic_discovery?) {
+    console.log("selected, new ->", leftClickSelectedNode, target);
+    var command = new AddLinkCommand(this.graphService.getGraph(), leftClickSelectedNode.getName(), target.getName(), timeout, circuit_breaker, dynamic_discovery);
+    this.graphInvoker.executeCommand(command);
+    this.navigation.getPaper().findViewByModel(leftClickSelectedNode).unhighlight();
+    console.log("added link");
   }
 
   reverseLink(link) {
@@ -162,6 +84,11 @@ export class ArchitectureEditingService {
     let target = link.target();
     link.source(target);
     link.target(source);
+  }
+
+  removeLink(link: joint.shapes.microtosca.RunTimeLink) {
+    console.log("removing link called");
+    this.graphInvoker.executeCommand(new RemoveLinkCommand(this.graphService.getGraph(), link));
   }
 
 }
